colocado uma verificação de null pra puxar os socios no service pra deletar

+++ src/main/java/com/omni/cadastro_cnpj/service/CnpjService.java
@@ -97,8 +97,10 @@ public class CnpjService implements ICnpjService {
     public void deletar(Long id) {
         Cnpj cnpjExistente = cnpjRepository.findById(id)
                 .orElseThrow(() -> new CnpjNotFoundException(id));
-        for(Socio socio : cnpjExistente.getSocios()) {
-        	socioRepository.delete(socio);
+        if (cnpjExistente.getSocios() != null) {
+	        for(Socio socio : cnpjExistente.getSocios()) {
+	        	socioRepository.delete(socio);
+	        }
         }
         cnpjRepository.delete(cnpjExistente);
     }
